@@ -6,7 +6,7 @@ description: Demonstrates how to use the media_and_create_date_picker plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -19,8 +19,8 @@ dependencies:
     # The example app is bundled with the plugin so we use a path dependency on
     # the parent directory to use the current plugin's version. 
     path: ../
-  heic_to_jpg: ^0.1.2
-  video_thumbnail: ^0.2.1
+  heic_to_jpg: ^0.2.0
+  video_thumbnail: ^0.3.3
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
