@@ -4,8 +4,8 @@ version: 0.2.2
 homepage: https://github.com/unbam/media_and_create_date_picker
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.10.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.20.0"
 
 dependencies:
   flutter:
